Switched to an in-memory DB for UnitTest
Can't figure it out what was causing the socket connection exception, so, since we will perform only basic operation, we can switch to an in-memory implementation of the DB for the unit tests. We will then try a real DB implementation in integration tests

Also disabling Integration-test for the moment since there are none and the CI building is failing.